init: use `write region` to prepend string
For whatever reason, use `find-file-noselect` here introduces a lot of
overheads, which make `find-file-noselect` takes almost forever to open a file
and return it as a buffer (seems that it only happens on emacs30+?).  To make it
worse, `save-buffer` is also expensive here.

Since all we wanted is to prepend a string to a file, `write-region` would be a
better option.  After this patch `rem -r` shell command should feel as smooth as
butter.